+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+### [1.6.5] - 2025-07-29
+### Improved
+- added display of popup messages during shipments registration via bulk action in the WC orders page
+
+### Updated
+- itella-api library to v2.4.7
+
 ### [1.6.4] - 2025-07-11
 ### Fixed
 - fixed the Sender ID of the courier invitation to use the plugin developer ID instead of the sender's company code